+++ tests/TelegramBotApiTest.php
@@ -53,6 +53,7 @@ use Phptg\BotApi\Type\Sticker\Sticker;
 use Phptg\BotApi\Type\Story;
 use Phptg\BotApi\Type\User;
 use Phptg\BotApi\Type\UserChatBoosts;
+use Phptg\BotApi\Type\UserProfileAudios;
 use Phptg\BotApi\Type\UserProfilePhotos;
 use Phptg\BotApi\Type\Update\Update;
 use Phptg\BotApi\Type\Update\WebhookInfo;
@@ -1518,6 +1519,18 @@ final class TelegramBotApiTest extends TestCase
         assertInstanceOf(OwnedGifts::class, $result);
     }
 
+    public function testGetUserProfileAudios(): void
+    {
+        $api = TestHelper::createSuccessStubApi([
+            'total_count' => 0,
+            'audios' => [],
+        ]);
+
+        $result = $api->getUserProfileAudios(7);
+
+        assertInstanceOf(UserProfileAudios::class, $result);
+    }
+
     public function testGetUserProfilePhotos(): void
     {
         $api = TestHelper::createSuccessStubApi([
@@ -1672,6 +1685,15 @@ final class TelegramBotApiTest extends TestCase
         $api = TestHelper::createSuccessStubApi(true);
 
         $result = $api->removeBusinessAccountProfilePhoto('connection1');
+
+        assertTrue($result);
+    }
+
+    public function testRemoveMyProfilePhoto(): void
+    {
+        $api = TestHelper::createSuccessStubApi(true);
+
+        $result = $api->removeMyProfilePhoto();
 
         assertTrue($result);
     }
@@ -2308,6 +2330,19 @@ final class TelegramBotApiTest extends TestCase
         $api = TestHelper::createSuccessStubApi(true);
 
         $result = $api->setMyName();
+
+        assertTrue($result);
+    }
+
+    public function testSetMyProfilePhoto(): void
+    {
+        $api = TestHelper::createSuccessStubApi(true);
+
+        $result = $api->setMyProfilePhoto(
+            new InputProfilePhotoStatic(
+                new InputFile((new StreamFactory())->createStream()),
+            ),
+        );
 
         assertTrue($result);
     }