+++ tests/Type/UserTest.php
@@ -32,6 +32,7 @@ final class UserTest extends TestCase
         assertNull($user->canConnectToBusiness);
         assertNull($user->hasMainWebApp);
         assertNull($user->hasTopicsEnabled);
+        assertNull($user->allowsUsersToCreateTopics);
     }
 
     public function testToRequestArray(): void
@@ -51,6 +52,7 @@ final class UserTest extends TestCase
             true,
             false,
             true,
+            true,
         );
 
         assertSame(
@@ -69,6 +71,7 @@ final class UserTest extends TestCase
                 'can_connect_to_business' => true,
                 'has_main_web_app' => false,
                 'has_topics_enabled' => true,
+                'allows_users_to_create_topics' => true,
             ],
             $user->toRequestArray(),
         );
@@ -91,6 +94,7 @@ final class UserTest extends TestCase
             'can_connect_to_business' => true,
             'has_main_web_app' => false,
             'has_topics_enabled' => true,
+            'allows_users_to_create_topics' => true,
         ], null, User::class);
 
         assertInstanceOf(User::class, $user);
@@ -108,5 +112,6 @@ final class UserTest extends TestCase
         assertSame(true, $user->canConnectToBusiness);
         assertSame(false, $user->hasMainWebApp);
         assertSame(true, $user->hasTopicsEnabled);
+        assertSame(true, $user->allowsUsersToCreateTopics);
     }
 }